% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biggerHdf5}
\alias{biggerHdf5}
\title{Find the bigger HDF5 file}
\usage{
biggerHdf5(where = "local")
}
\arguments{
\item{where}{"local" for project. The other option is "package"}
}
\description{
Find the bigger HDF5 file
}
\keyword{internal}
