% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfactor.R
\name{calcCriticals}
\alias{calcCriticals}
\title{Calculate the pressure and temperature criticals}
\usage{
calcCriticals(pres.a, temp.f, gas.sg, co2.frac = 0, h2s.frac = 0,
  n2.frac = 0)
}
\arguments{
\item{pres.a}{absolute pressure, psia}

\item{temp.f}{temperature, deg F}

\item{gas.sg}{gas specific gravity}

\item{co2.frac}{CO2}

\item{h2s.frac}{H2S}

\item{n2.frac}{N2}
}
\value{
list
}
\description{
Returns a list of calculated pseudo-reduced and pseudo-critical
pressure and temperature
}
