% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copyDataContainer}
\alias{copyDataContainer}
\title{Copy a HDF5 data container to user's folder}
\usage{
copyDataContainer(overwrite = FALSE)
}
\arguments{
\item{overwrite}{it is FALSE to prevent overwriting an existing data file}
}
\description{
It will check if a HDF5 file already exist at the destination.
If the file exist it will throw a warning.
To overwrite it with a fresh copy use overwrite=TRUE
}
