% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNodal-package.r
\docType{data}
\name{curves.0241}
\alias{curves.0241}
\title{Digitized version of Fancher and Brown curve. Brown Chapter 2 figure 41}
\format{An object of class \code{data.frame} with 30 rows and 3 columns.}
\usage{
data(curves.0241)
}
\description{
Digitized version of Fancher and Brown curve. Brown Chapter 2 figure 41
}
\keyword{datasets}
