% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duns-Ros.R
\name{dunsros.0}
\alias{dunsros.0}
\title{Duns and Ros correlation}
\usage{
dunsros.0(pres, temp, surf.params)
}
\arguments{
\item{pres}{pressure                  psia}

\item{temp}{temperature               deg Far}

\item{surf.params}{surface parameters list}
}
\description{
Duns and Ros correlation
}
