% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duns-Ros.R
\name{dunsros.holdup}
\alias{dunsros.holdup}
\title{Holdup function according to Duns & Ros
Equation in 2.62 pg 43 Ovadia Shoham
Equation 2.39, 2.310, 2.311, 2.312 in Brown pg 113}
\usage{
dunsros.holdup(vsl, liq.sg, if.tens, vsg, diam, liq.visc, pres)
}
\arguments{
\item{vsl}{liquid superficial velocity        ft/s}

\item{liq.sg}{liquid specif gravity}

\item{if.tens}{interfacial tension

TODO: pass liq.dens instead of liq.sg
TODO: move pres to start of params list}

\item{vsg}{superficial velocity of gas        ft/s}

\item{diam}{diameter of tubing                 ft}

\item{liq.visc}{liquid viscosity                   cp}

\item{pres}{pressure                           psia}
}
\description{
Holdup function according to Duns & Ros
Equation in 2.62 pg 43 Ovadia Shoham
Equation 2.39, 2.310, 2.311, 2.312 in Brown pg 113
}
