% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duns-Ros.R
\name{dunsros.numbers}
\alias{dunsros.numbers}
\title{Duns and Ros adimensional numbers}
\usage{
dunsros.numbers(pres, vsl, vsg, liq.dens, liq.visc, liq.surft, diam)
}
\arguments{
\item{pres}{pressure                           psia}

\item{vsl}{liquid superficial velocity        ft/s}

\item{vsg}{superficial velocity of gas        ft/s}

\item{liq.dens}{liquid density}

\item{liq.visc}{liquid viscosity                   cp}

\item{liq.surft}{liquid surface tension}

\item{diam}{diameter of tubing                 ft}
}
\description{
Duns and Ros adimensional numbers
}
