% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FANBR.R
\name{fanbr.fanbr}
\alias{fanbr.fanbr}
\title{Fancher and Brown correlation}
\usage{
fanbr.fanbr(pres, temp, surf.params)
}
\arguments{
\item{pres}{pressure                           psia}

\item{temp}{temperature                        degF}

\item{surf.params}{surface parameters list}
}
\description{
Fancher and Brown correlation
}
