% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moody.R
\name{ff.chen}
\alias{ff.chen}
\title{Chen Friction factor correl1ation
Guo pg 4/48}
\usage{
ff.chen(ed, Re)
}
\arguments{
\item{ed}{relative roughness}

\item{Re}{Reynolds number}
}
\description{
Chen Friction factor correl1ation
Guo pg 4/48
}
