% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileInfoHdf5}
\alias{fileInfoHdf5}
\title{File info for all HDF5 files}
\usage{
fileInfoHdf5(where = "local")
}
\arguments{
\item{where}{"local" for project. The other option is "package"}
}
\description{
File info for all HDF5 files
}
\keyword{internal}
