% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moody.R
\name{friction.factor}
\alias{friction.factor}
\title{Friction factor calculatiobn using Colebrook}
\usage{
friction.factor(REY, ED, f1 = blasius.0)
}
\arguments{
\item{REY}{Reynolds number}

\item{ED}{Relative roughness}

\item{f1}{a Blasius function}
}
\description{
Friction factor calculatiobn using Colebrook
}
