% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLUIDPROPS.R
\name{gas.fvf}
\alias{gas.fvf}
\title{Gas Formation Volume Factor, Bg.
After Craft, pg 23
Specify the units: cuft/SCF    "cuft"
                   bbl/SCF     "bbl"}
\usage{
gas.fvf(pres, temp, z, units = "cuft")
}
\arguments{
\item{pres}{pressure                 psia}

\item{temp}{temperature              deg Far}

\item{z}{compressibility factor}

\item{units}{units for volume}
}
\description{
Gas Formation Volume Factor, Bg.
After Craft, pg 23
Specify the units: cuft/SCF    "cuft"
                   bbl/SCF     "bbl"
}
