% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gascorrs.R
\name{gas.viscBase}
\alias{gas.viscBase}
\title{Correlation for gas base viscosity
at atmospheric conditions. Dempsey (1965)
Guo, pg. 2/23 Eq 2.36}
\usage{
gas.viscBase(temp, sgg)
}
\arguments{
\item{temp}{temperature deg F}

\item{sgg}{specific gravity of gas}
}
\description{
Correlation for gas base viscosity
at atmospheric conditions. Dempsey (1965)
Guo, pg. 2/23 Eq 2.36
}
