% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{get.group.paths}
\alias{get.group.paths}
\title{Get a vector of unique groups in HDF5 file}
\usage{
get.group.paths(hFile)
}
\arguments{
\item{hFile}{a HDF5 file to inspect}
}
\value{
a vector with unique full path groups
}
\description{
Get a vector of unique groups in HDF5 file
}
\keyword{internal}
