% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{get.well.cases}
\alias{get.well.cases}
\title{Get well observations. dataframe and count per /field/well}
\usage{
get.well.cases(hFile, fieldName, wellName)
}
\arguments{
\item{hFile}{name of the HDF5 file}

\item{fieldName}{field name}

\item{wellName}{well name}
}
\description{
Get well observations. dataframe and count per /field/well
}
