% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rutils.R
\name{get.well.slot}
\alias{get.well.slot}
\title{Add a record to HDF5 file of well inputs and outputs}
\usage{
get.well.slot(hFile, fieldName, wellName)
}
\arguments{
\item{hFile}{HDF5 file, full path}

\item{fieldName}{the field name}

\item{wellName}{the well name}
}
\value{
a string with /field/well/datetime if creation successful
          otherwise, NULL
}
\description{
Add a record to HDF5 file of well inputs and outputs
}
