% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{get_well_parameters}
\alias{get_well_parameters}
\title{Get the well input together with the basic calculations}
\usage{
get_well_parameters(well.input)
}
\arguments{
\item{well.input}{the well input}
}
\description{
Get the well input together with the basic calculations
}
