% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{getBasicCalcs}
\alias{getBasicCalcs}
\title{High level function that calls any VLP model}
\usage{
getBasicCalcs(well.input)
}
\arguments{
\item{well.input}{well input data as a list}
}
\description{
High level function that calls any VLP model
}
