% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDefaultDataContainerName}
\alias{getDefaultDataContainerName}
\title{Get the name of the default DataContainer}
\usage{
getDefaultDataContainerName()
}
\description{
Get the name of the default DataContainer
}
\keyword{internal}
