% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGBR.R
\name{hagbr.dummy}
\alias{hagbr.dummy}
\title{Hagedorn-Brown dummy function using 1/log10(P)
and 1/log10(T) to calculate dp.dz}
\usage{
hagbr.dummy(pres, temp, surf.params)
}
\arguments{
\item{pres}{pressure}

\item{temp}{temperature degF}

\item{surf.params}{a list of surface parameters}
}
\description{
Hagedorn-Brown dummy function using 1/log10(P)
and 1/log10(T) to calculate dp.dz
}
