% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGBR.R
\name{hagbr.guo}
\alias{hagbr.guo}
\title{Hagedorn-Brown correlation as shown in Guo, Lyons, Ghalambor book
Beware that this correlation does not make P, T corrections for
fluid properties with exception of z factor}
\usage{
hagbr.guo(pres, temp, surf.params)
}
\arguments{
\item{pres}{pressure}

\item{temp}{temperature degF}

\item{surf.params}{a list of surface parameters}
}
\description{
Hagedorn-Brown correlation as shown in Guo, Lyons, Ghalambor book
Beware that this correlation does not make P, T corrections for
fluid properties with exception of z factor
}
