% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGBR.R
\name{hagbr.mod}
\alias{hagbr.mod}
\title{Hagedorn correlation from the Brown's book. Procedure C.42}
\usage{
hagbr.mod(pres, temp, well.params)
}
\arguments{
\item{pres}{pressure at depth          psia    dbl}

\item{temp}{temperature at depth       deg F   dbl}

\item{surf.params}{surface parameters            list}
}
\description{
Hagedorn correlation from the Brown's book. Procedure C.42
}
