% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{hGroup.exists}
\alias{hGroup.exists}
\title{Check if a HDF5 specific group exists already in the file}
\usage{
hGroup.exists(hFile, hGroup)
}
\arguments{
\item{hFile}{a full path to a HDF5 file}

\item{hGroup}{any group}
}
\value{
TRUE if group already exists
          FALSE if group has not been created yet
}
\description{
Check if a HDF5 specific group exists already in the file
}
\keyword{internal}
