% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duns-Ros.R
\name{holdup.griffith}
\alias{holdup.griffith}
\title{Griffith correlation for bubble-flow regime
Mentioned in Brown 2.3332 Generalized correlation of HagBr}
\usage{
holdup.griffith(vsm, vsg, vsl = 0.8)
}
\arguments{
\item{vsm}{superficial velocity of mixture}

\item{vsg}{superficial velocity of gas}

\item{vsl}{superficial velocity of liquid}
}
\description{
Griffith correlation for bubble-flow regime
Mentioned in Brown 2.3332 Generalized correlation of HagBr
}
