% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_hdf5_files}
\alias{is_hdf5_files}
\title{Logical response to presence of HDF5 files anywhere under user root folder}
\usage{
is_hdf5_files(where = "local")
}
\description{
Logical response to presence of HDF5 files anywhere under user root folder
}
\keyword{internal}
