% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{loadVLP}
\alias{loadVLP}
\title{Load only the source necessary for model or correlation
Note: it doesn't unload the functions sourced yet}
\usage{
loadVLP(model)
}
\arguments{
\item{model}{the model name}
}
\description{
Load only the source necessary for model or correlation
Note: it doesn't unload the functions sourced yet
}
