% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friction.R
\name{moody.ff}
\alias{moody.ff}
\title{Moody friction factor calculation}
\usage{
moody.ff(Re, ed, tol = 1e-05, trials = 100)
}
\arguments{
\item{Re}{Reynolds number}

\item{ed}{Relative roughness}

\item{tol}{tolerance}

\item{trials}{number of calculations to try}
}
\description{
Moody friction factor calculation
}
