% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gascorrs.R
\name{mu.gas}
\alias{mu.gas}
\title{Calculate gas viscosity
Guo, pg 2/23}
\usage{
mu.gas(pres, temp, gas.sg)
}
\arguments{
\item{pres}{pressure}

\item{temp}{temperature}

\item{gas.sg}{gas specific gavity}
}
\description{
Calculate gas viscosity
Guo, pg 2/23
}
