% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gascorrs.R
\name{mu.r}
\alias{mu.r}
\title{Reduced viscosity of gas
Dempsey correlation to find the viscosity of gas
Guo, pg 2/23}
\usage{
mu.r(pres.pr, temp.pr)
}
\arguments{
\item{pres.pr}{pseudo reduced pressure}

\item{temp.pr}{pseudo reduced temperature}
}
\description{
Reduced viscosity of gas
Dempsey correlation to find the viscosity of gas
Guo, pg 2/23
}
