% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named.list}
\alias{named.list}
\title{Take objects and create a list using their names}
\usage{
named.list(...)
}
\arguments{
\item{...}{any additional parameters}
}
\description{
This avoids retyping the name of the object as in a list
}
\keyword{internal}
