% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLUIDPROPS.R
\name{oil.dead_visc.BeggsRobinson}
\alias{oil.dead_visc.BeggsRobinson}
\title{Dead Oil viscosity by Beggs-Robinson}
\usage{
oil.dead_visc.BeggsRobinson(temp, API)
}
\arguments{
\item{temp}{temperature}

\item{API}{oil specific gravity}
}
\description{
Dead Oil viscosity by Beggs-Robinson
}
