% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLUIDPROPS.R
\name{oil.fvf.glaso}
\alias{oil.fvf.glaso}
\title{Formation Volume Factor by Glaso}
\usage{
oil.fvf.glaso(temp, Rs, oil.sg, gas.sg)
}
\arguments{
\item{temp}{temperature}

\item{Rs}{solution ratio}

\item{oil.sg}{oil specific gravity}

\item{gas.sg}{gas specific gravity}
}
\description{
Formation Volume Factor by Glaso
}
