% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLUIDPROPS.R
\name{oil.fvf}
\alias{oil.fvf}
\title{Oil Formation Volumme Factor}
\usage{
oil.fvf(pres, temp, oil.sg)
}
\arguments{
\item{pres}{pressure}

\item{temp}{temperature}

\item{oil.sg}{oil specific gravity}
}
\description{
Oil Formation Volumme Factor
}
