% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil_correlations.R
\name{oil.fvf.standing}
\alias{oil.fvf.standing}
\title{Oil Formation Volume Factor. Saturated. Standing}
\usage{
oil.fvf.standing(temp, Rs, oil.sg, gas.sg)
}
\arguments{
\item{temp}{temperature}

\item{Rs}{gas oil solution ratio}

\item{oil.sg}{oil specific gravity}

\item{gas.sg}{gas specific gravity}
}
\description{
Oil Formation Volume Factor. Saturated. Standing
}
