% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil_correlations.R
\name{oil.Rs.standing}
\alias{oil.Rs.standing}
\title{Gas oil solution ratio}
\usage{
oil.Rs.standing(pres, temp, API, gas.sg)
}
\arguments{
\item{pres}{pressure}

\item{temp}{temperature}

\item{API}{oil specific gravity}

\item{gas.sg}{gas specific gravity}
}
\description{
Gas oil solution ratio
}
