% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil_correlations.R
\name{oil.sat_visc.BeggsRobinson}
\alias{oil.sat_visc.BeggsRobinson}
\title{Saturated oil viscosity by Beggs-Robinson
Valid for pressures below the bubble point}
\usage{
oil.sat_visc.BeggsRobinson(temp, API, Rs)
}
\arguments{
\item{temp}{temperature}

\item{API}{oil density}

\item{Rs}{gas oil solution ratio}
}
\description{
Saturated oil viscosity by Beggs-Robinson
Valid for pressures below the bubble point
}
