% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLUIDPROPS.R
\name{oil.shrinkage}
\alias{oil.shrinkage}
\title{Oil Shrinkage
Craft, pg 35}
\usage{
oil.shrinkage(oil.fvf)
}
\arguments{
\item{oil.fvf}{Oil Formation Volume factor         bbl/stb}
}
\description{
Oil Shrinkage
Craft, pg 35
}
