% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{readFromProjectEnv}
\alias{readFromProjectEnv}
\title{Read variable from environment}
\usage{
readFromProjectEnv(var)
}
\arguments{
\item{var}{variable to read}
}
\description{
Read variable from environment
}
\keyword{internal}
