% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{runVLP}
\alias{runVLP}
\title{Run the VLP calculations}
\usage{
runVLP(well.input, model.parameters)
}
\arguments{
\item{well.input}{well input data as a list}

\item{model.parameters}{well model parameters as a list}
}
\description{
Run the VLP calculations
}
