% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{saveToProjectEnv}
\alias{saveToProjectEnv}
\title{Save a variable to the project environment}
\usage{
saveToProjectEnv(var, value)
}
\arguments{
\item{var}{variable to save in environment}

\item{value}{what is to be saved}
}
\description{
Save a variable to the project environment
}
\keyword{internal}
