% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{set_deviation_survey}
\alias{set_deviation_survey}
\title{Build the deviation survey table given MD and TVD}
\usage{
set_deviation_survey(well_as_string)
}
\arguments{
\item{well_as_string}{well deviation survey as two columns text}
}
\description{
Build the deviation survey table given MD and TVD
}
