% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/class-VLP.R
\docType{methods}
\name{set_vlp_input}
\alias{set_vlp_input}
\alias{set_vlp_input,VLP-method}
\title{Set the VLP input for the well}
\usage{
set_vlp_input(object, ...)

\S4method{set_vlp_input}{VLP}(object, field.name = "HAGBR.GUO",
  well.name = "ExampleGuo", depth.wh = 0, depth.bh = 9700,
  diam.in = 1.995, GLR = 75, liq.rt = 758, wcut = 0.1, thp = 200,
  tht = 80, bht = 180, API = 40, gas.sg = 0.7, wat.sg = 1.05,
  oil.visc = 5, ed = 6e-04, if.tens = 30, salinity = 0, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}

\item{field.name}{field name. Usually comprises several wells}

\item{well.name}{well name}

\item{depth.wh}{depth at wellhead}

\item{depth.bh}{depth at bottomhole of well                    feet}

\item{diam.in}{well diameter                                  inches}

\item{GLR}{gas liquid ratio}

\item{liq.rt}{liquid rate}

\item{wcut}{watercut}

\item{thp}{tubing head pressure}

\item{tht}{tubing head temperature}

\item{bht}{bottomhole temperature}

\item{API}{gravity of oil}

\item{gas.sg}{specific gravity of gas}

\item{wat.sg}{specific gravity of water}

\item{oil.visc}{oil viscosity}

\item{ed}{relative rougness}

\item{if.tens}{interfacial tension between ...}

\item{salinity}{water salinity}
}
\description{
Set the VLP input for the well

Set the VLP inputs
}
