% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/class-VLP.R
\docType{methods}
\name{set_vlp_model}
\alias{set_vlp_model}
\alias{set_vlp_model,VLP-method}
\title{Set the VLP model definition}
\usage{
set_vlp_model(object, ...)

\S4method{set_vlp_model}{VLP}(object, vlp.model = "hagbr.guo",
  segments = 29, tol = 1e-04, dp.dz.ini = 0.002, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}

\item{vlp.model}{the name of the VLP model or correlation}

\item{segments}{number of segments for the tubing string}

\item{tol}{tolerance for error during interations}

\item{dp.dz.ini}{initial gradient}
}
\description{
such as tolerance for iterations, initial value of dp/dz, the HDF5 data file
}
