% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{setVLPmodel}
\alias{setVLPmodel}
\title{Set model parameters such as tolerance for iterations, initial value of dp/dz
the HDF5 data file}
\usage{
setVLPmodel(vlp.model = "hagbr.guo", segments = 29, tol = 1e-04,
  dp.dz.ini = 0.002)
}
\arguments{
\item{vlp.model}{the name of the VLP model or correlation}

\item{segments}{number of segments for the tubing string}

\item{tol}{tolerance for error during interations}

\item{dp.dz.ini}{initial gradient}
}
\description{
Set model parameters such as tolerance for iterations, initial value of dp/dz
the HDF5 data file
}
