% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VLP.R
\docType{class}
\name{VLP-class}
\alias{VLP-class}
\title{VLP class}
\arguments{
\item{object}{a class object}

\item{well.parameters}{an ODE object}

\item{model.parameters}{an ODE object}

\item{vlpInput}{the VLP input}

\item{vlpModel}{the VLP model input}

\item{...}{additional parameters}
}
\description{
VLP class
}
