% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VLP.R
\name{VLP}
\alias{VLP}
\title{VLP constructor}
\usage{
VLP(well.parameters, model.parameters)
}
\arguments{
\item{well.parameters}{well production input}

\item{model.parameters}{the VLP model data}
}
\description{
VLP constructor
}
