% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{VLPcontrol}
\alias{VLPcontrol}
\title{VLP calculation algorithm}
\usage{
VLPcontrol(well.parameters, model.parameters, verbose = FALSE)
}
\arguments{
\item{well.parameters}{well input and core calculations at surface}

\item{model.parameters}{model characteristics. Hagedorn-Brown, Duns-Ros,
Fancher-Brown, etc. Also tolerances and boundaries}
}
\description{
VLP control marching algorithm to calculate bottomhole pressure given the
wellhead pressure and temperature. Fluid properties are calculated at each
of the segment depths
}
