% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLUIDPROPS.R
\name{water.fvf.mccoy}
\alias{water.fvf.mccoy}
\title{Water Formation Volume Factor for water saturated wit gas. McCoy
From book Correlaciones Numericas PVT by Banzer pg 118}
\usage{
water.fvf.mccoy(pres, temp, salinity = 0)
}
\arguments{
\item{pres}{pressure                                  psia}

\item{temp}{temperature                               deg F}

\item{salinity}{salinity of water                         ppm}
}
\description{
Water Formation Volume Factor for water saturated wit gas. McCoy
From book Correlaciones Numericas PVT by Banzer pg 118
}
