% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/class-well_deviation_survey.R
\docType{class}
\name{WellDeviationSurvey}
\alias{WellDeviationSurvey}
\alias{WellDeviationSurvey-class}
\alias{.WellDeviationSurvey}
\alias{WellDeviationSurvey,character-method}
\alias{WellDeviationSurvey,missing-method}
\title{WellDeviationSurvey structure}
\usage{
WellDeviationSurvey(md_tvd, ...)

\S4method{WellDeviationSurvey}{character}(md_tvd, ...)

\S4method{WellDeviationSurvey}{missing}(md_tvd, ...)
}
\arguments{
\item{md_tvd}{MD and TVD}

\item{...}{additional parameters}
}
\description{
WellDeviationSurvey structure
}
