% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-well_deviation_survey.R
\name{compute_angle_deviation_survey}
\alias{compute_angle_deviation_survey}
\title{Compute the angle of a deviated well and cumulative displacement}
\usage{
compute_angle_deviation_survey(well_table, reference = c("vertical",
  "horizontal"))
}
\arguments{
\item{well_table}{a well deviation survey}

\item{reference}{angle against the vertical or horizontal}
}
\description{
Compute the angle of a deviated well and cumulative displacement
}
