% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLP.R
\name{writeHdf5}
\alias{writeHdf5}
\title{Write dataset to HDF5 file}
\usage{
writeHdf5(dataTable, dataset.name)
}
\arguments{
\item{dataTable}{data that is going to be written}

\item{dataset.name}{the name of the dataset}
}
\description{
Write dataset to HDF5 file
}
\keyword{internal}
