% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfactor.R
\name{z.brillbeggs}
\alias{z.brillbeggs}
\title{Calculate the Z factor with the Brill-Beggs correlation}
\usage{
z.brillbeggs(pres.a, temp.f, gas.sg, n2.frac = 0, co2.frac = 0,
  h2s.frac = 0)
}
\arguments{
\item{pres.a}{absolute pressure, psia}

\item{temp.f}{temperature, deg F}

\item{gas.sg}{gas specific gravity}

\item{n2.frac}{N2}

\item{co2.frac}{CO2}

\item{h2s.frac}{H2S}
}
\description{
Calculate the Z factor with the Brill-Beggs correlation
}
\section{Gas correlations}{
NA
}

\seealso{
Other Z factor correlations: \code{\link{z.hallyarborough}},
  \code{\link{zcorr}}
}
