% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfactor.R
\name{z.hallyarborough}
\alias{z.hallyarborough}
\title{Calculate the Z factor using the Hall-Yarborough method}
\usage{
z.hallyarborough(pres.a, temp.f, gas.sg, n2.frac = 0, co2.frac = 0,
  h2s.frac = 0, as_list = FALSE)
}
\arguments{
\item{pres.a}{absolute pressure, psia}

\item{temp.f}{temperature, deg F}

\item{gas.sg}{gas specific gravity}

\item{n2.frac}{N2}

\item{co2.frac}{CO2}

\item{h2s.frac}{H2S}

\item{as_list}{FALSE to return single value z; otherwise returns a list}
}
\description{
Calculate the Z factor using the Hall-Yarborough method
}
\section{Gas correlations}{
NA
}

\seealso{
Other Z factor correlations: \code{\link{z.brillbeggs}},
  \code{\link{zcorr}}
}
