% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfactor.R
\name{z.hallyarborough}
\alias{z.hallyarborough}
\alias{calcCriticals}
\alias{z.brillbeggs}
\title{Calculate Z factor using Hall-Yarborough method}
\usage{
z.hallyarborough(pres.a, temp.f, gas.sg, n2.frac = 0, co2.frac = 0,
  h2s.frac = 0)

calcCriticals(pres.a, temp.f, gas.sg, co2.frac = 0, h2s.frac = 0,
  n2.frac = 0)

z.brillbeggs(pabs, tempFar, gas.sg, n2.frac, co2.frac, h2s.frac)
}
\arguments{
\item{pres.a}{absolute pressure, psia}

\item{temp.f}{temperature, deg F}

\item{gas.sg}{gas specific gravity}

\item{n2.frac}{N2}

\item{co2.frac}{CO2}

\item{h2s.frac}{H2S}

\item{pabs}{absolute pressure, psia}

\item{tempFar}{temperature degF}
}
\description{
Calculate Z factor using Hall-Yarborough method

Calculate the pressure and temperature criticals
Returns list
}
