% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpt}
\alias{dpt}
\title{dataset from Guo, example calculations}
\format{An object of class \code{data.frame} with 30 rows and 3 columns.}
\usage{
data(dpt)
}
\description{
dataset from Guo, example calculations
}
\keyword{datasets}
