% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil_correlations.R
\name{Rs.glaso}
\alias{Rs.glaso}
\title{Glaso}
\usage{
Rs.glaso(pres, temp, API, gas.sg)
}
\arguments{
\item{pres}{pressure}

\item{temp}{temperature}

\item{API}{oil density}

\item{gas.sg}{gas specific gravity}
}
\description{
Glaso
}
