% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VLP.R
\docType{class}
\name{VLP2-class}
\alias{VLP2-class}
\title{VLP class}
\arguments{
\item{well_input}{an ODE object}

\item{vlp_model}{an ODE object}

\item{...}{additional parameters}
}
\description{
VLP class
}
